Removed ID token parsing
Removed code that gets user's email address from ID token. This code was
problematic for a few reasons:

- It required addition of the `profile` scope.
- It assumed that the `preferred_username` claim in the ID token was an
SMTP address. This is not always the case (e.g. user may sign in with a
mobile phone number)

Replaced the `getEmailFromIdToken` method with a new `getUserEmail`
method.

Note that this requires an updated version of node_outlook (v1.1.6 or
above).